Don't cast away provenance in ptr2int casts
When casting from a pointer to an integer that can hold a pointer (e.g.,
uint64_t in the current code base), cast to uintptr_t first and allow an
implicit cast to make the final cast rather than doing:

 a_uint64 = (uint64_t)(uintptr_t)a_pointer;

This will avoid loss of provenance on CHERI when the destination is
updated to be a type capable of holding a capability.

+++ lib/libzfs/libzfs_util.c
@@ -1189,7 +1189,7 @@ zcmd_alloc_dst_nvlist(libzfs_handle_t *hdl, zfs_cmd_t *zc, size_t len)
 		len = 256 * 1024;
 	zc->zc_nvlist_dst_size = len;
 	zc->zc_nvlist_dst =
-	    (uint64_t)(uintptr_t)zfs_alloc(hdl, zc->zc_nvlist_dst_size);
+	    (uintptr_t)zfs_alloc(hdl, zc->zc_nvlist_dst_size);
 }
 
 /*
@@ -1202,7 +1202,7 @@ zcmd_expand_dst_nvlist(libzfs_handle_t *hdl, zfs_cmd_t *zc)
 {
 	free((void *)(uintptr_t)zc->zc_nvlist_dst);
 	zc->zc_nvlist_dst =
-	    (uint64_t)(uintptr_t)zfs_alloc(hdl, zc->zc_nvlist_dst_size);
+	    (uintptr_t)zfs_alloc(hdl, zc->zc_nvlist_dst_size);
 }
 
 /*
@@ -1230,7 +1230,7 @@ zcmd_write_nvlist_com(libzfs_handle_t *hdl, uint64_t *outnv, uint64_t *outlen,
 
 	verify(nvlist_pack(nvl, &packed, &len, NV_ENCODE_NATIVE, 0) == 0);
 
-	*outnv = (uint64_t)(uintptr_t)packed;
+	*outnv = (uintptr_t)packed;
 	*outlen = len;
 }
